More app delegate clean up
