--- Client/Application/AppDelegate+Experiments.swift
@@ -1,53 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0
-
-import Shared
-
-private let log = Logger.browserLogger
-
-import Foundation
-
-extension AppDelegate {
-    func initializeExperiments() {
-        // We intialize the generated FxNimbus singleton very early on with a lazily
-        // constructed singleton.
-        FxNimbus.shared.initialize(with: { Experiments.shared })
-        // We also make sure that any cache invalidation happens after each applyPendingExperiments().
-        NotificationCenter.default.addObserver(forName: .nimbusExperimentsApplied, object: nil, queue: nil) { _ in
-            FxNimbus.shared.invalidateCachedValues()
-        }
-
-        let defaults = UserDefaults.standard
-        let nimbusFirstRun = "NimbusFirstRun"
-        let isFirstRun = defaults.object(forKey: nimbusFirstRun) == nil
-        defaults.set(false, forKey: nimbusFirstRun)
-        Experiments.customTargetingAttributes =  ["isFirstRun": "\(isFirstRun)"]
-        let initialExperiments = Bundle.main.url(forResource: "initial_experiments", withExtension: "json")
-        let serverURL = Experiments.remoteSettingsURL
-        let savedOptions = Experiments.getLocalExperimentData()
-        let options: Experiments.InitializationOptions
-        switch (savedOptions, isFirstRun, initialExperiments, serverURL) {
-        // QA testing case: experiments come from the Experiments setting screen.
-        case (let payload, _, _, _) where payload != nil:
-            log.info("Nimbus: Loading from experiments provided by settings screen")
-            options = Experiments.InitializationOptions.testing(localPayload: payload!)
-        // First startup case:
-        case (nil, true, let file, _) where file != nil:
-            log.info("Nimbus: Loading from experiments from bundle, at first startup")
-            options = Experiments.InitializationOptions.preload(fileUrl: file!)
-        // Local development case: load from the bundled initial_experiments.json
-        case (_, _, let file, let url) where file != nil && url == nil:
-            log.info("Nimbus: Loading from experiments from bundle, with no URL")
-            options = Experiments.InitializationOptions.preload(fileUrl: file!)
-        case (_, _, _, let url) where url != nil:
-            log.info("Nimbus: server exists")
-            options = Experiments.InitializationOptions.normal
-        default:
-            log.info("Nimbus: server does not exist")
-            options = Experiments.InitializationOptions.normal
-        }
-
-        Experiments.intialize(options)
-    }
-}
